Move ProvenanceIR to types to merge with ValidatedProvenance
