+++ internal/verifier/verifier.go
@@ -128,7 +128,7 @@ func chdir(dir string) {
 // all non-empty fields in got using fields in the reference values. Empty fields will not be verified.
 // TODO(#69): Check metadata against the expected values.
 type ProvenanceIRVerifier struct {
-	Got  *common.ProvenanceIR
+	Got  *types.ProvenanceIR
 	Want *common.ReferenceValues
 }
 
@@ -172,7 +172,7 @@ func (verifier *ProvenanceIRVerifier) Verify() (VerificationResult, error) {
 }
 
 // verifyBinarySHA256Digest verifies that the binary SHA256 in this provenance is contained in the given reference binary SHA256 digests (in want).
-func verifyBinarySHA256Digest(want *common.ReferenceValues, got *common.ProvenanceIR) (VerificationResult, error) {
+func verifyBinarySHA256Digest(want *common.ReferenceValues, got *types.ProvenanceIR) (VerificationResult, error) {
 	result := NewVerificationResult()
 
 	gotBinarySHA256Digest, err := got.GetBinarySHA256Digest()
@@ -203,7 +203,7 @@ func verifyBinarySHA256Digest(want *common.ReferenceValues, got *common.Provenan
 }
 
 // verifyHasBuildCmd verifies that the build cmd is not empty.
-func verifyHasBuildCmd(got *common.ProvenanceIR) VerificationResult {
+func verifyHasBuildCmd(got *types.ProvenanceIR) VerificationResult {
 	result := NewVerificationResult()
 	if _, err := got.GetBuildCmd(); err != nil {
 		result.SetFailed("no build cmd found")
@@ -212,7 +212,7 @@ func verifyHasBuildCmd(got *common.ProvenanceIR) VerificationResult {
 }
 
 // verifyBuilderImageDigest verifies that the given builder image digest matches a builder image digest in the reference values.
-func verifyBuilderImageDigest(want *common.ReferenceValues, got *common.ProvenanceIR) (VerificationResult, error) {
+func verifyBuilderImageDigest(want *common.ReferenceValues, got *types.ProvenanceIR) (VerificationResult, error) {
 	result := NewVerificationResult()
 
 	gotBuilderImageDigest, err := got.GetBuilderImageSHA256Digest()
@@ -238,7 +238,7 @@ func verifyBuilderImageDigest(want *common.ReferenceValues, got *common.Provenan
 }
 
 // verifyRepoURIs verifies that the references to URIs in the provenance point to the repo URI given in the reference values.
-func verifyRepoURIs(want *common.ReferenceValues, got *common.ProvenanceIR) VerificationResult {
+func verifyRepoURIs(want *common.ReferenceValues, got *types.ProvenanceIR) VerificationResult {
 	result := NewVerificationResult()
 
 	for _, gotRepoURI := range got.GetRepoURIs() {
